+++ jenkins-client/src/main/java/com/offbytwo/jenkins/client/JenkinsHttpClient.java
@@ -3,7 +3,6 @@
  *
  * Distributed under the MIT license: http://opensource.org/licenses/MIT
  */
-
 package com.offbytwo.jenkins.client;
 
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -19,6 +18,8 @@ import com.offbytwo.jenkins.model.ExtractHeader;
 import net.sf.json.JSONObject;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.StringUtils;
+import org.apache.http.Header;
+import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
 import org.apache.http.NameValuePair;
 import org.apache.http.auth.AuthScope;
@@ -30,6 +31,9 @@ import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.methods.HttpRequestBase;
 import org.apache.http.entity.ContentType;
 import org.apache.http.entity.StringEntity;
+import org.apache.http.entity.mime.HttpMultipartMode;
+import org.apache.http.entity.mime.MultipartEntityBuilder;
+import org.apache.http.entity.mime.content.FileBody;
 import org.apache.http.impl.auth.BasicScheme;
 import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.apache.http.impl.client.CloseableHttpClient;
@@ -41,6 +45,7 @@ import org.apache.http.util.EntityUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
@@ -50,12 +55,10 @@ import java.util.Map;
 import static com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES;
 import com.offbytwo.jenkins.client.util.ResponseUtils;
 import com.offbytwo.jenkins.client.util.UrlUtils;
-import java.io.Closeable;
 import static org.apache.commons.lang.StringUtils.isNotBlank;
 
-//import com.offbytwo.jenkins.client.util.HttpResponseContentExtractor;
+public class JenkinsHttpClient implements JenkinsHttpConnection {
 
-public class JenkinsHttpClient implements Closeable {
     private final Logger LOGGER = LoggerFactory.getLogger(getClass());
 
     private URI uri;
@@ -70,7 +73,7 @@ public class JenkinsHttpClient implements Closeable {
     private String jenkinsVersion;
 
     public final static String EMPTY_VERSION = "UNKNOWN";
-    
+
     /**
      * Create an unauthenticated Jenkins HTTP client
      *
@@ -138,14 +141,9 @@ public class JenkinsHttpClient implements Closeable {
     }
 
     /**
-     * Perform a GET request and parse the response to the given class
-     *
-     * @param path path to request, can be relative or absolute
-     * @param cls class of the response
-     * @param <T> type of the response
-     * @return an instance of the supplied class
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public <T extends BaseModel> T get(String path, Class<T> cls) throws IOException {
         HttpGet getMethod = new HttpGet(UrlUtils.toJsonApiUri(uri, context, path));
 
@@ -170,31 +168,32 @@ public class JenkinsHttpClient implements Closeable {
 	 * @param <T>
 	 *            type of the response
 	 * @return an List of instances of the supplied class
-	 * @throws IOException
-	 *             in case of an error.
 	 */
-	public <T extends BaseModel> List<T> get(String path, TypeReference<List<T>> list) throws IOException {
+	public <T extends BaseModel> List<T> get(String path, TypeReference<List<T>> list) {
 		HttpGet getMethod = new HttpGet(UrlUtils.toJsonApiUri(uri, context, path));
-
-		HttpResponse response = client.execute(getMethod, localContext);
-		jenkinsVersion = ResponseUtils.getJenkinsVersion(response);
+		HttpResponse response = null;
 		try {
+			response = client.execute(getMethod, localContext);
+			jenkinsVersion = ResponseUtils.getJenkinsVersion(response);
 			httpResponseValidator.validateResponse(response);
 			return objectFromResponse(list, response);
+		} catch (IOException e) {
+			new RuntimeException(e); 
 		} finally {
-			EntityUtils.consume(response.getEntity());
+			try {
+				EntityUtils.consume(response.getEntity());
+			} catch (IOException e) {
+				new RuntimeException(e); 
+			}
 			releaseConnection(getMethod);
 		}
+		return null;
 	}
 
     /**
-     * Perform a GET request and parse the response and return a simple string
-     * of the content
-     *
-     * @param path path to request, can be relative or absolute
-     * @return the entity text
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public String get(String path) throws IOException {
         HttpGet getMethod = new HttpGet(UrlUtils.toJsonApiUri(uri, context, path));
         HttpResponse response = client.execute(getMethod, localContext);
@@ -212,14 +211,9 @@ public class JenkinsHttpClient implements Closeable {
     }
 
     /**
-     * Perform a GET request and parse the response to the given class, logging
-     * any IOException that is thrown rather than propagating it.
-     *
-     * @param path path to request, can be relative or absolute
-     * @param cls class of the response
-     * @param <T> type of the response
-     * @return an instance of the supplied class
+     * {@inheritDoc}
      */
+    @Override
     public <T extends BaseModel> T getQuietly(String path, Class<T> cls) {
         T value;
         try {
@@ -233,12 +227,9 @@ public class JenkinsHttpClient implements Closeable {
     }
 
     /**
-     * Perform a GET request and return the response as InputStream
-     *
-     * @param path path to request, can be relative or absolute
-     * @return the response stream
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public InputStream getFile(URI path) throws IOException {
         HttpGet getMethod = new HttpGet(path);
         HttpResponse response = client.execute(getMethod, localContext);
@@ -247,23 +238,27 @@ public class JenkinsHttpClient implements Closeable {
         return new RequestReleasingInputStream(response.getEntity().getContent(), getMethod);
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public <R extends BaseModel, D> R post(String path, D data, Class<R> cls) throws IOException {
-        return post(path, data, cls, true);
+        return post(path, data, cls, null, true);
     }
 
     /**
-     * Perform a POST request and parse the response to the given class
-     *
-     * @param path path to request, can be relative or absolute
-     * @param data data to post
-     * @param cls class of the response
-     * @param <R> type of the response
-     * @param <D> type of the data
-     * @param crumbFlag true / false.
-     * @return an instance of the supplied class
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public <R extends BaseModel, D> R post(String path, D data, Class<R> cls, boolean crumbFlag) throws IOException {
+        return post(path, data, cls, null, crumbFlag);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public <R extends BaseModel, D> R post(String path, D data, Class<R> cls, Map<String, File> fileParams, boolean crumbFlag) throws IOException {
         HttpPost request = new HttpPost(UrlUtils.toJsonApiUri(uri, context, path));
         if (crumbFlag == true) {
             Crumb crumb = getQuietly("/crumbIssuer", Crumb.class);
@@ -277,6 +272,21 @@ public class JenkinsHttpClient implements Closeable {
             StringEntity stringEntity = new StringEntity(value, ContentType.APPLICATION_JSON);
             request.setEntity(stringEntity);
         }
+
+        // Prepare file parameters
+        if(fileParams != null && !(fileParams.isEmpty())) {
+            MultipartEntityBuilder builder = MultipartEntityBuilder.create();
+            builder.setMode(HttpMultipartMode.BROWSER_COMPATIBLE);
+
+            for (Map.Entry<String, File> entry : fileParams.entrySet()) {
+                FileBody fileBody = new FileBody(entry.getValue());
+                builder.addPart(entry.getKey(), fileBody);
+            }
+
+            HttpEntity entity = builder.build();
+            request.setEntity(entity);
+        }
+
         HttpResponse response = client.execute(request, localContext);
         jenkinsVersion = ResponseUtils.getJenkinsVersion(response);
 
@@ -303,23 +313,9 @@ public class JenkinsHttpClient implements Closeable {
     }
 
     /**
-     * Perform a POST request using form url encoding.
-     * 
-     * This method was added for the purposes of creating folders, but may be
- useful for other API calls as well.
- 
- Unlike post and post_xml, the path is *not* modified by adding
- "/toJsonApiUri/json". Additionally, the params in data are provided as both
- request parameters including a json parameter, *and* in the
- JSON-formatted StringEntity, because this is what the folder creation
- call required. It is unclear if any other jenkins APIs operate in this
- fashion.
-     *
-     * @param path path to request, can be relative or absolute
-     * @param data data to post
-     * @param crumbFlag true / false.
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public void post_form(String path, Map<String, String> data, boolean crumbFlag) throws IOException {
         HttpPost request;
         if (data != null) {
@@ -357,18 +353,10 @@ public class JenkinsHttpClient implements Closeable {
         }
     }
 
-
     /**
-     * Perform a POST request using form url encoding and return HttpResponse object
-     * This method is not performing validation and can be used for more generic queries to jenkins.
-     *
-     * @param path
-     *            path to request, can be relative or absolute
-     * @param data
-     *            data to post
-     * @throws IOException,
-     *             HttpResponseException
+     * {@inheritDoc}
      */
+    @Override
     public HttpResponse post_form_with_result(String path, List<NameValuePair> data, boolean crumbFlag) throws IOException {
         HttpPost request;
         if (data != null) {
@@ -391,18 +379,17 @@ public class JenkinsHttpClient implements Closeable {
     }
 
     /**
-     * Perform a POST request of XML (instead of using json mapper) and return a
-     * string rendering of the response entity.
-     *
-     * @param path path to request, can be relative or absolute
-     * @param xml_data data data to post
-     * @return A string containing the xml response (if present)
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public String post_xml(String path, String xml_data) throws IOException {
         return post_xml(path, xml_data, true);
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public String post_xml(String path, String xml_data, boolean crumbFlag) throws IOException {
         HttpPost request = new HttpPost(UrlUtils.toJsonApiUri(uri, context, path));
         if (crumbFlag == true) {
@@ -427,28 +414,17 @@ public class JenkinsHttpClient implements Closeable {
     }
 
     /**
-     * Post a text entity to the given URL using the default content type
-     *
-     * @param path The path.
-     * @param textData data.
-     * @param crumbFlag true/false.
-     * @return resulting response
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public String post_text(String path, String textData, boolean crumbFlag) throws IOException {
         return post_text(path, textData, ContentType.DEFAULT_TEXT, crumbFlag);
     }
 
     /**
-     * Post a text entity to the given URL with the given content type
-     *
-     * @param path The path.
-     * @param textData The data.
-     * @param contentType {@link ContentType}
-     * @param crumbFlag true or false.
-     * @return resulting response
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public String post_text(String path, String textData, ContentType contentType, boolean crumbFlag)
             throws IOException {
         HttpPost request = new HttpPost(UrlUtils.toJsonApiUri(uri, context, path));
@@ -474,17 +450,19 @@ public class JenkinsHttpClient implements Closeable {
     }
 
     /**
-     * Perform POST request that takes no parameters and returns no response
-     *
-     * @param path path to request
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public void post(String path) throws IOException {
-        post(path, null, null, false);
+        post(path, null, null, null, false);
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public void post(String path, boolean crumbFlag) throws IOException {
-        post(path, null, null, crumbFlag);
+        post(path, null, null, null, crumbFlag);
     }
 
     
@@ -501,7 +479,82 @@ public class JenkinsHttpClient implements Closeable {
 	}
 
     
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String getJenkinsVersion() {
+        return this.jenkinsVersion;
+    }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean isJenkinsVersionSet() {
+        return !EMPTY_VERSION.equals(this.jenkinsVersion);
+    }
+
+    /**
+     * Closes underlying resources. Any I/O errors whilst closing are logged
+     * with debug log level Closed instances should no longer be used Closing an
+     * already closed instance has no side effects
+     */
+    @Override
+    public void close() {
+        try {
+            client.close();
+        } catch (final IOException ex) {
+            LOGGER.debug("I/O exception closing client", ex);
+        }
+    }
+
+    
+    /**
+     * Add authentication to supplied builder.
+     * @param builder the builder to configure
+     * @param uri the server URI
+     * @param username the username
+     * @param password the password
+     * @return the passed in builder
+     */
+    protected static HttpClientBuilder addAuthentication(final HttpClientBuilder builder, 
+            final URI uri, final String username,
+            String password) {
+        if (isNotBlank(username)) {
+            CredentialsProvider provider = new BasicCredentialsProvider();
+            AuthScope scope = new AuthScope(uri.getHost(), uri.getPort(), "realm");
+            UsernamePasswordCredentials credentials = new UsernamePasswordCredentials(username, password);
+            provider.setCredentials(scope, credentials);
+            builder.setDefaultCredentialsProvider(provider);
+
+            builder.addInterceptorFirst(new PreemptiveAuth());
+        }
+        return builder;
+    }
+
+    
+    /**
+     * Get the local context.
+     * @return context
+     */
+    protected HttpContext getLocalContext() {
+        return localContext;
+    }
+
+    
+    /**
+     * Set the local context.
+     * @param localContext the context
+     */
+    protected void setLocalContext(final HttpContext localContext) {
+        this.localContext = localContext;
+    }
+
+    
+    
+    
+    
     private <T extends BaseModel> T objectFromResponse(Class<T> cls, HttpResponse response) throws IOException {
         InputStream content = response.getEntity().getContent();
         byte[] bytes = ByteStreams.toByteArray(content);
@@ -518,70 +571,8 @@ public class JenkinsHttpClient implements Closeable {
         return mapper;
     }
 
-    /**
-     * @return the version string.
-     */
-    public String getJenkinsVersion() {
-        return this.jenkinsVersion;
-    }
-
-    /**
-     * Check to see if the jenkins version has been set
-     * to something different than the initialization value
-     * from the constructor. This means there has never been made
-     * a communication with the Jenkins server.
-     * @return true if jenkinsVersion has been set by communication, false otherwise. 
-     */
-    public boolean isJenkinsVersionSet() {
-        return !EMPTY_VERSION.equals( this.jenkinsVersion );
-    }
-    
-    
-    
-    
-    /**
-     * Closes underlying resources.
-     * Any I/O errors whilst closing are logged with debug log level
-     * Closed instances should no longer be used
-     * Closing an already closed instance has no side effects
-     */
-    @Override
-    public void close() {
-        try {
-            client.close();
-        } catch (final IOException ex) {
-            LOGGER.debug("I/O exception closing client", ex);
-        }        
-    }
-    
-
-    
-    
     private void releaseConnection(HttpRequestBase httpRequestBase) {
         httpRequestBase.releaseConnection();
     }
 
-    protected static HttpClientBuilder addAuthentication(HttpClientBuilder builder, URI uri, String username,
-            String password) {
-        if (isNotBlank(username)) {
-            CredentialsProvider provider = new BasicCredentialsProvider();
-            AuthScope scope = new AuthScope(uri.getHost(), uri.getPort(), "realm");
-            UsernamePasswordCredentials credentials = new UsernamePasswordCredentials(username, password);
-            provider.setCredentials(scope, credentials);
-            builder.setDefaultCredentialsProvider(provider);
-
-            builder.addInterceptorFirst(new PreemptiveAuth());
-        }
-        return builder;
-    }
-
-    protected HttpContext getLocalContext() {
-        return localContext;
-    }
-
-    protected void setLocalContext(HttpContext localContext) {
-        this.localContext = localContext;
-    }
-
-    
 }